# payment-service/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Payment API
    version: 0.0.1
paths:
    v1/create:
        post:
            tags:
                - Payment
            operationId: Payment_CreatePayment
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CreateResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    v1/delete/{paymentId}:
        delete:
            tags:
                - Payment
            operationId: Payment_DeletePayment
            parameters:
                - name: paymentId
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/DeleteRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeleteResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    v1/find/{paymentId}:
        get:
            tags:
                - Payment
            operationId: Payment_FindPayment
            parameters:
                - name: paymentId
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/FindRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/FindResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    v1/update:
        put:
            tags:
                - Payment
            operationId: Payment_UpdatePayment
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UpdateResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
components:
    schemas:
        CreateRequest:
            type: object
            properties:
                paymentId:
                    type: string
                fromAccountId:
                    type: string
                toAccountId:
                    type: string
                paymentType:
                    type: string
                amount:
                    type: string
                currency:
                    type: string
                paymentMethod:
                    type: string
                referenceNumber:
                    type: string
                status:
                    type: string
        CreateResponse:
            type: object
            properties:
                paymentId:
                    type: string
                fromAccountId:
                    type: string
                toAccountId:
                    type: string
                paymentType:
                    type: string
                amount:
                    type: string
                currency:
                    type: string
                paymentMethod:
                    type: string
                referenceNumber:
                    type: string
                status:
                    type: string
        DeleteRequest:
            type: object
            properties:
                paymentId:
                    type: string
                fromAccountId:
                    type: string
                toAccountId:
                    type: string
                paymentType:
                    type: string
                amount:
                    type: string
                currency:
                    type: string
                paymentMethod:
                    type: string
                referenceNumber:
                    type: string
                status:
                    type: string
        DeleteResponse:
            type: object
            properties:
                paymentId:
                    type: string
                fromAccountId:
                    type: string
                toAccountId:
                    type: string
                paymentType:
                    type: string
                amount:
                    type: string
                currency:
                    type: string
                paymentMethod:
                    type: string
                referenceNumber:
                    type: string
                status:
                    type: string
        FindRequest:
            type: object
            properties:
                paymentId:
                    type: string
                fromAccountId:
                    type: string
                toAccountId:
                    type: string
                paymentType:
                    type: string
                amount:
                    type: string
                currency:
                    type: string
                paymentMethod:
                    type: string
                referenceNumber:
                    type: string
                status:
                    type: string
        FindResponse:
            type: object
            properties:
                paymentId:
                    type: string
                fromAccountId:
                    type: string
                toAccountId:
                    type: string
                paymentType:
                    type: string
                amount:
                    type: string
                currency:
                    type: string
                paymentMethod:
                    type: string
                referenceNumber:
                    type: string
                status:
                    type: string
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
        UpdateRequest:
            type: object
            properties:
                paymentId:
                    type: string
                fromAccountId:
                    type: string
                toAccountId:
                    type: string
                paymentType:
                    type: string
                amount:
                    type: string
                currency:
                    type: string
                paymentMethod:
                    type: string
                referenceNumber:
                    type: string
                status:
                    type: string
        UpdateResponse:
            type: object
            properties:
                paymentId:
                    type: string
                fromAccountId:
                    type: string
                toAccountId:
                    type: string
                paymentType:
                    type: string
                amount:
                    type: string
                currency:
                    type: string
                paymentMethod:
                    type: string
                referenceNumber:
                    type: string
                status:
                    type: string
tags:
    - name: Payment